Extract FormNavGuard to its own component
